when "querying all" avoid seeing delisted days for published events
when deleted, we have to show the delisted days otherwise the data for the event will be missing.
but we don't want to show delisted days when the event is still valid.
the difference can be seen by changing a caldaily eventstatus to "D" for an otherwise valid calevent. that individual day shouldn't be visible. ( and it was before this change )