fix cal wordwrapping  #709
the wordwrapping utility we're using trims lines by default after it wraps them. needed to turn that off.

now the ics lines are output as:
```
DESCRIPTION:Portland might be famous for its MAX trains and Tillikum
 Crossing these days\, but back at the turn of the 20th Century Portland\,
```
hard to see -- but now there is a space after "Tillikum " ( and after "Portalnd\, " ) where previously there wasn't. the space in front of " Crossing" was there before, and is required for proper line handling in ical.

added an option to the url (filename=none) to render the results as plain text -- helps a lot for debugging.

https://github.com/75lb/wordwrapjs?tab=readme-ov-file#module_wordwrapjs--Wordwrap..WordwrapOptions