normalize usage of docker-compose
this change makes it so that macos uses the same compose command as production.

this is a "patch" until we can upgrade production. once that's done, then this should be converted to "docker compose" for all platforms. ( the "dash" version is deprecated, but the "space" version doesn't yet exist on production  )

i've tested this on apple silicon with local Docker version 26.1.1, build 4cf5afa. and i've tested this previously on my older intel mac with ... a relatively recent version of docker.