move environment config to docker .env
prior to this change the ./shift script was *required* to run docker, because the script sets the default environment variables.
that makes running on windows complicated.
moving the environment to docker's .env makes ./shift optional.
it's still used /needed on production ( esp for ./shift pull )
and macos and linux users _should_ use it.
but windows users  can use "docker compose" directly, or the new npm up/down