fix scrolling content
the disclaimer footer and the event listings were separately srcollable. to fix this, had to parent the router/content and the toolbar panel content to the same element. which meant making the toolbar *just* the buttons.
CalMain therefore says what buttons to show, and then chooses what to show when active