move tool bar to top; give it internal scrolling
stop the scroll pos of the events list from being disrupted when the toolbar panel opens and closes. accomplished by making the top and bottom fixed to the edges.
that requires the content to be a fixed size too. gave the main page an overall structure "top", "mid", "bottom" for that. for the scrolling panels, gave panels its own grouping under mid.
added "sticky" dividers for the dates... because why not.
saves scroll position when going _to_ the details, that way favorites scroll position gets saved too.