fix bug with favorite icon not updating
when you're on a page and click the favorite button its state wasn't changing. this was a regression introduced when making the prev and next arrows grey out at the right times.
now using a "watcher" to look for/rebuild shortcut button changes. a bit obscure but seems to work okay for both cases.